Create Product.jsx
created a useState for the img to initially display when the page loads and will hopefully display when the an img is not being hovered over or clicked. I made a side gallery section for the images. I wrote some pseudocode to get. a better idea of what I want the side gallery to do.